requirements.txt
prometheus-client==0.12.0
pyaml==21.10.1
Flask~=2.2.3
jsonpath-python~=1.0.6
